+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This file is used to list changes made in each version of hyper-kube-config.
 
 ## unreleased
 
+## 0.1.0 (2019-05-14)
+- [Zane Williamson] 
+  - Adding support to store and associate Pems to a cluster
+  - Adding support to set and get arbitrary status and environment strings to a cluster
+
 ## 0.0.3 (2018-04-01)
 - [Akshay Raj] - Fixed fixed YAMLLoadWarning